requirements.txt
os
python-dotenv
google-generativeai
pymysql
streamlit
pandas